fix(loki): address remaining endpoint regressions
