+++ src/sinks/loki/config.rs
@@ -3,6 +3,7 @@ use std::collections::HashMap;
 use vrl::value::Kind;
 
 use super::{
+    append_loki_path,
     healthcheck::healthcheck,
     sink::{LokiSink, confine_template_keys},
 };
@@ -23,26 +24,6 @@ const fn default_compression() -> Compression {
 
 fn default_loki_path() -> String {
     "/loki/api/v1/push".to_string()
-}
-
-/// Appends Loki's push path while retaining the boundary normalization used by
-/// `UriSerde::append_path`: all trailing slashes on the endpoint and all
-/// leading slashes on the configured path collapse to one separator.
-fn append_loki_path(endpoint: &HttpEndpoint, path: &str) -> crate::Result<HttpEndpoint> {
-    let mut parts = endpoint.as_uri().clone().into_parts();
-    let base_path = parts
-        .path_and_query
-        .as_ref()
-        .map(http::uri::PathAndQuery::path)
-        .unwrap_or_default();
-    let joined = format!(
-        "{}/{}",
-        base_path.trim_end_matches('/'),
-        path.trim_start_matches('/')
-    );
-    parts.path_and_query = Some(joined.parse()?);
-
-    Ok(HttpEndpoint::new(http::Uri::from_parts(parts)?)?)
 }
 
 /// Configuration for the `loki` sink.
@@ -403,7 +384,7 @@ pub fn valid_label_name(label: &Template) -> bool {
 mod tests {
     use std::convert::TryInto;
 
-    use super::{append_loki_path, valid_label_name};
+    use super::{super::append_loki_path, valid_label_name};
     use crate::{
         config::ValidatedSink,
         sinks::{loki::LokiConfig, util::HttpEndpoint},
